+++ lua/obsidian/commands/backlinks.lua
@@ -7,7 +7,7 @@ return function(client, _)
   ---@type obsidian.Note|?
   local note
   local cursor_link, _, ref_type = util.parse_cursor_link()
-  if cursor_link ~= nil and ref_type ~= RefTypes.NakedUrl then
+  if cursor_link ~= nil and ref_type ~= RefTypes.NakedUrl and ref_type ~= RefTypes.FileUrl then
     note = client:resolve_note(cursor_link)
     if note == nil then
       log.err "Could not resolve link under cursor to a note ID, path, or alias"